--- run_no_out.sh
@@ -1,1 +1,0 @@
-nohup python3 dc_bot_main.py
